fix: Transaction表單與api定義的interface不一樣

+++ src/views/Records.vue
@@ -88,7 +88,9 @@
 					message: t('records.confirm_delete_message'),
 				})
 					.then(() => {
-						return transactionApi.deleteTransaction(item.id)
+						if (item.id) {
+							return transactionApi.deleteTransaction(item.id)
+						}
 					})
 					.then(() => {
 						waterfallRef.value?.refresh()
